refactor(common): date 유틸 common으로 이동 및 헬퍼 함수 추가
- date.util을 src/utils에서 src/common으로 이동 (utils/date.util.ts 삭제)
- addMinutes: 시작 시간 + 이용 시간(분) → 종료 시간 Date 반환
- toKSTTimeStr: Date → KST HH:mm 문자열 반환
- 예약 수정 성공 메시지 시작~종료 시간 형식으로 변경
